Add bookings (#2)
* add-bookings
добавлены БД, jpa Рефакторинг кода под новое состояние.
добавлена возможность бронирования и комментирования.

+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.service.UserService;
 import ru.practicum.shareit.validateGroups.Create;
 import ru.practicum.shareit.validateGroups.Update;
 
@@ -22,7 +23,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public UserDto getUserById(@PathVariable int id) {
+    public UserDto getUserById(@PathVariable Long id) {
         return userService.getUsersById(id);
     }
 
@@ -32,12 +33,12 @@ public class UserController {
     }
 
     @PatchMapping("/{id}")
-    public UserDto update(@PathVariable int id, @Validated(Update.class) @RequestBody UserDto userDto) {
+    public UserDto update(@PathVariable Long id, @Validated(Update.class) @RequestBody UserDto userDto) {
         return userService.update(id, userDto);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable Long id) {
         userService.delete(id);
     }
 }
